Move GamePrototype to its own file
